app/backend/agents/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
langgraph>=0.0.26
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-aws>=0.0.1
pydantic-settings>=2.1.0
structlog>=24.1.0
motor>=3.3.2
boto3>=1.34.0
pydantic>=2.6.0
python-dotenv>=1.0.1
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation-fastapi>=0.43b0
langdetect>=1.0.9
langchain-aws
lingua-language-detector
pytz
